refactor(configuration): add unit test and refactor
