+++ packages/elements/src/configuration/helpers/types.ts
@@ -1,0 +1,9 @@
+export type IconMap = {
+  [key: string]: string,
+}
+export type ConfigIcon = {
+  map: IconMap
+}
+export type Config = {
+  icon: ConfigIcon,
+}
