+++ packages/elements/src/configuration/index.ts
@@ -1,2 +1,3 @@
 export * from './elements/configuration.js';
-export { efConfig, type Config } from './helpers/context.js';
+export { efConfig } from './helpers/context.js';
+export type { Config } from './helpers/types.js';
